% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_explore_activity}
\alias{run_explore_activity}
\title{Interactive tool to explore code activity}
\usage{
run_explore_activity()
}
\description{
This function launches the Shiny app for visualizing the data in this package.
}
