% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snomed_dict}
\alias{snomed_dict}
\title{Dictionary of SNOMED Code Descriptions in Primary Care in England}
\format{
\subsection{\code{snomed_dict}}{

A data frame with 1,366,513 rows and 4 columns:
\describe{
\item{start_date}{Start date of code description}
\item{end_date}{End date of code description}
\item{snomed_concept_id}{SNOMED Concept ID}
\item{description}{Description of SNOMED Concept ID}
}
}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/mi-snomed-code-usage-in-primary-care}
}
\usage{
snomed_dict
}
\description{
Dictionary of code descriptions (including SNOMED concepts) for each year.
Note that there are sometimes small differences across years.
}
\keyword{datasets}
